Remove nonsense defaults from auth config
For some reason, Past Gentlehacker thought it made sense to allow users to include default values for many of the parameters in authentication requests. This is, in fact, not very helpful; If these parameters are required they need to be sent by the client and, if not, can safely be skipped.

This change removes defaults for all authentication request parameters except for the `client_id`.